Try only with buildings for now
